+++ src/State.cpp
@@ -1023,7 +1023,7 @@ std::vector<std::pair<std::string, std::string>>* State::GetDefines()
 bool State::ShaderEnabled(const RE::BSShader::Type a_type)
 {
 	auto index = magic_enum::enum_integer(a_type) + 1;
-	if (index < sizeof(enabledClasses)) {
+	if (index < static_cast<int>(sizeof(enabledClasses))) {
 		return enabledClasses[index];
 	}
 	return false;
